# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CGPanel Administrator and Management API",
    "version": "0.3.0",
    "description": "Community alpha. Session + CSRF or administrator bearer tokens. See docs/API.md for operational boundaries.",
    "license": {
      "name": "AGPL-3.0-or-later",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Your panel HTTPS origin, normally port 2083"
    }
  ],
  "paths": {
...
    "/api/logout": {
      "post": {
        "operationId": "post_api_logout",
        "summary": "End browser session",
        "tags": [
          "Authentication"
        ],
        "description": "Session and CSRF only. Does not revoke API tokens.",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Job submissions return queued, not completion.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Application errors usually return an error object. Framework/proxy errors may differ. See API guide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              },
              "example": {}
            }
          }
        }
      }
    },
    "/api/password": {
      "post": {
        "operationId": "post_api_password",
        "summary": "Change own password",
        "tags": [
          "Authentication"
        ],
        "description": "Session and CSRF only. Requires current password; new password 14–256 characters. Revokes all own sessions and API tokens.",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Job submissions return queued, not completion.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Application errors usually return an error object. Framework/proxy errors may differ. See API guide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "current": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "current",
                  "password"
                ]
              },
              "example": {
                "current": "CURRENT_PASSWORD",
                "password": "NEW_PASSWORD"
              }
            }
          }
        }
      }
    },
    "/api/admin/system": {
      "get": {
        "operationId": "get_api_admin_system",
        "summary": "Administrator system overview",
        "tags": [
          "Administration"
        ],
        "description": "Administrator only. Version, resource counts, recent events and broker host health.",
        "security": [
          {
            "AdminBearer": []
          },
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Job submissions return queued, not completion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/System"
                }
              }
            }
          },
          "default": {
            "description": "Application errors usually return an error object. Framework/proxy errors may differ. See API guide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/audit": {
      "get": {
        "operationId": "get_api_admin_audit",
        "summary": "Paginated audit events",
        "tags": [
          "Administration"
        ],
        "description": "Administrator only. Events in descending ID order. Pass next_before as before until null. Includes token IDs for bearer mutations.",
        "security": [
          {
            "AdminBearer": []
          },
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Job submissions return queued, not completion.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "next_before": {
                      "type": "integer",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Application errors usually return an error object. Framework/proxy errors may differ. See API guide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "1–200, defaults to 50",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "before",
            "in": "query",
            "description": "Exclusive positive event ID cursor",
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/api/admin/tokens": {
      "get": {
        "operationId": "get_api_admin_tokens",
        "summary": "List own API token metadata",
        "tags": [
          "Administration"
        ],
        "description": "Administrator session only. Latest 200 records; no secret or hash. Expired and revoked records remain visible.",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Job submissions return queued, not completion.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TokenMetadata"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Application errors usually return an error object. Framework/proxy errors may differ. See API guide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post_api_admin_tokens",
        "summary": "Create administrator API token",
        "tags": [
          "Administration"
        ],
        "description": "Administrator session and CSRF only. Secret shown once. At most 20 active tokens, 1–90 days; account and token CIDRs both apply. Defaults: read scope, 30 days, no extra IP filter.",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Job submissions return queued, not completion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenSecret"
                }
              }
            }
          },
          "default": {
            "description": "Application errors usually return an error object. Framework/proxy errors may differ. See API guide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenCreate"
              },
              "example": {
                "name": "reporting",
                "scope": "read",
                "expires_days": 30,
                "allowed_ips": [
                  "203.0.113.10/32"
                ]
              }
            }
          }
        }
      }
    },
    "/api/admin/tokens/{id}": {
      "delete": {
        "operationId": "delete_api_admin_tokens_id",
        "summary": "Revoke own API token",
        "tags": [
          "Administration"
        ],
        "description": "Administrator session and CSRF only. Idempotent for an existing own token. Does not cancel in-flight requests or queued jobs.",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Job submissions return queued, not completion.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Application errors usually return an error object. Framework/proxy errors may differ. See API guide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/admin/openapi.json": {
      "get": {
        "operationId": "get_api_admin_openapi_json",
        "summary": "Download OpenAPI contract",
        "tags": [
          "Administration"
        ],
        "description": "Administrator only. Same contract checked into docs/openapi.json.",
        "security": [
          {
            "AdminBearer": []
          },
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Job submissions return queued, not completion.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Application errors usually return an error object. Framework/proxy errors may differ. See API guide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Revocable cgp_ administrator token. read scope permits GET/HEAD only; admin permits mutations except session-only credential operations."
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cg_session"
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-csrf-token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "TokenCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "scope": {
            "type": "string",
            "enum": [
              "read",
              "admin"
            ]
          },
          "expires_days": {
            "type": "integer",
            "minimum": 1,
            "maximum": 90,
            "default": 30
          },
          "allowed_ips": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "TokenMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": "string"
          },
          "scope": {
            "type": "string",
            "enum": [
              "read",
              "admin"
            ]
          },
          "allowed_ips": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created": {
            "type": "integer"
          },
          "expires": {
            "type": "integer"
          },
          "last_used": {
            "type": "integer",
            "nullable": true
          },
          "revoked": {
            "type": "integer",
            "nullable": true
          }
        }
      },
      "TokenSecret": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TokenMetadata"
          },
          {
            "type": "object",
            "properties": {
              "token": {
                "type": "string",
                "description": "Returned only on creation. Store privately."
              }
            },
            "required": [
              "token"
            ]
          }
        ]
      },
...
      "System": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string"
          },
          "counts": {
            "type": "object",
            "additionalProperties": true
          },
          "events": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "host": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    }
  },
...
}